[Fixed]1.完善消息也得数据处理 2.通话页临时密码可以复制

+++ app/src/main/java/com/philips/easykey/lock/bean/SevendayDataStatisticsBean.java
@@ -6,7 +6,7 @@ public class SevendayDataStatisticsBean {
 
     private String statisticsTypeName;
 
-    private float[] ordinateValue ;
+    private int[] ordinateValue ;
 
     private String[] transverseValue;
 
@@ -18,11 +18,11 @@ public class SevendayDataStatisticsBean {
         this.statisticsTypeName = statisticsTypeName;
     }
 
-    public float[] getOrdinateValue() {
+    public int[] getOrdinateValue() {
         return ordinateValue;
     }
 
-    public void setOrdinateValue(float[] ordinateValue) {
+    public void setOrdinateValue(int[] ordinateValue) {
         this.ordinateValue = ordinateValue;
     }
